requirements.txt
Flask==0.10.1
Flask-Cors==1.3.1
Flask-Migrate==1.2.0
Flask-RESTful==0.2.12
Flask-SQLAlchemy==1.0
Flask-Script==0.6.7
Flask-Testing==0.4.1
Jinja2==2.7.2
Mako==0.9.1
MarkupSafe==0.19
PyYAML==3.11
Pygments==1.6
SQLAlchemy==0.9.4
Sphinx==1.2.2
Werkzeug==0.9.4
alembic==0.6.4
amqp==1.4.4
aniso8601==0.82
anyjson==0.3.3
backports.ssl-match-hostname==3.4.0.2
billiard==3.3.0.16
celery==3.1.10
docutils==0.11
flower==0.6.0
gevent==1.0
greenlet==0.4.2
gunicorn==18.0
honcho==0.5.0
itsdangerous==0.24
kombu==3.0.14
nose==1.3.1
psycopg2==2.5.2
python-dateutil==2.2
pytz==2014.2
requests==2.2.1
six==1.6.1
tornado==3.2
wsgiref==0.1.2
